Removed all traces of old brute force code from sph
