+++ src/Headers/SphNeighbourSearch.h
@@ -113,122 +113,6 @@ protected:
 };
 
 
-
-//=================================================================================================
-//  Class SphBruteForceSearch
-/// \brief   Class for computing SPH neighbour lists using brute force only.
-/// \details Class for computing SPH neighbour lists using brute force only
-///          (i.e. direct summation over all particles).
-/// \author  D. A. Hubber, G. Rosotti
-/// \date    03/04/2013
-//=================================================================================================
-template <int ndim, template<int> class ParticleType>
-class SphBruteForceSearch : public SphNeighbourSearch<ndim>, public BruteForceSearch<ndim,ParticleType>
-{
- public:
-
-  using NeighbourSearch<ndim>::neibcheck;
-  using NeighbourSearch<ndim>::timing;
-  using NeighbourSearch<ndim>::kernp;
-  using NeighbourSearch<ndim>::kernfac;
-  using NeighbourSearch<ndim>::kernrange;
-  using NeighbourSearch<ndim>::kernrangesqd;
-
-
-  //-----------------------------------------------------------------------------------------------
-  SphBruteForceSearch(FLOAT, DomainBox<ndim> *, SmoothingKernel<ndim> *, CodeTiming *);
-  virtual ~SphBruteForceSearch();
-
-
-  //-----------------------------------------------------------------------------------------------
-  void UpdateAllSphProperties(int, int, SphParticle<ndim> *, Sph<ndim> *, Nbody<ndim> *);
-  void UpdateAllSphForces(int, int, SphParticle<ndim> *, Sph<ndim> *,
-                          Nbody<ndim> *, DomainBox<ndim> &, Ewald<ndim> *);
-  void UpdateAllSphHydroForces(int, int, SphParticle<ndim> *, Sph<ndim> *,
-                               Nbody<ndim> *, DomainBox<ndim> &);
-  void UpdateAllSphGravForces(int, int, SphParticle<ndim> *, Sph<ndim> *,
-                              Nbody<ndim> *, DomainBox<ndim> &, Ewald<ndim> *);
-  //void UpdateAllStarGasForces(int, int, SphParticle<ndim> *, Sph<ndim> *, Nbody<ndim> *);
-
-
-  GhostNeighbourFinder<ndim> GhostFinder; ///< Creates ghosts on the fly
-
-/*#ifdef MPI_PARALLEL
-  using NeighbourSearch<ndim>::ids_active_particles;
-  void UpdateGravityExportList(int, int, int, SphParticle<ndim> *, Sph<ndim> *, Nbody<ndim> *);
-  void UpdateHydroExportList(int, int, int, SphParticle<ndim> *,  Sph<ndim> *, Nbody<ndim> *);
-#endif*/
-
-};
-
-
-
-//=================================================================================================
-//  Class GradhSphBruteForce
-/// \brief   Class for computing neighbour lists using brute force only for grad-h SPH.
-/// \details Class for computing SPH neighbour lists using brute force only
-///          (i.e. direct summation over all particles) for grad-h SPH.
-/// \author  D. A. Hubber, G. Rosotti
-/// \date    12/05/2014
-//=================================================================================================
-template <int ndim, template<int> class ParticleType>
-class GradhSphBruteForce: public SphBruteForceSearch<ndim,ParticleType>
-{
- public:
-
-  using SphNeighbourSearch<ndim>::neibcheck;
-  using SphNeighbourSearch<ndim>::timing;
-  using SphNeighbourSearch<ndim>::kernp;
-  using SphNeighbourSearch<ndim>::kernfac;
-  using SphNeighbourSearch<ndim>::kernrange;
-  using SphNeighbourSearch<ndim>::kernrangesqd;
-
-
-  //-----------------------------------------------------------------------------------------------
-  GradhSphBruteForce(FLOAT, DomainBox<ndim> *, SmoothingKernel<ndim> *, CodeTiming *);
-  virtual ~GradhSphBruteForce();
-
-
-  //-----------------------------------------------------------------------------------------------
-  void UpdateAllSphProperties(int, int, SphParticle<ndim> *, Sph<ndim> *, Nbody<ndim> *);
-
-};
-
-
-
-//=================================================================================================
-//  Class SM2012SphBruteForce
-/// \brief   Class for computing neighbour lists using brute force only for SM2012 SPH.
-/// \details Class for computing neighbour lists using brute force only
-///          (i.e. direct summation over all particles) for SM2012 SPH.
-/// \author  D. A. Hubber, G. Rosotti
-/// \date    12/05/2014
-//=================================================================================================
-template <int ndim, template<int> class ParticleType>
-class SM2012SphBruteForce: public SphBruteForceSearch<ndim,ParticleType>
-{
- public:
-
-  using SphNeighbourSearch<ndim>::neibcheck;
-  using SphNeighbourSearch<ndim>::timing;
-  using SphNeighbourSearch<ndim>::kernp;
-  using SphNeighbourSearch<ndim>::kernfac;
-  using SphNeighbourSearch<ndim>::kernrange;
-  using SphNeighbourSearch<ndim>::kernrangesqd;
-
-
-  //-----------------------------------------------------------------------------------------------
-  SM2012SphBruteForce(FLOAT, DomainBox<ndim> *, SmoothingKernel<ndim> *, CodeTiming *);
-  virtual ~SM2012SphBruteForce();
-
-
-  //-----------------------------------------------------------------------------------------------
-  void UpdateAllSphProperties(int, int, SphParticle<ndim> *, Sph<ndim> *, Nbody<ndim> *);
-
-};
-
-
-
 //=================================================================================================
 //  Class SphTree
 /// \brief   Class containing tree for efficient SPH neighbour searching and gravity calculations.
@@ -412,7 +296,8 @@ class SM2012SphTree: public SphTree<ndim,ParticleType,TreeCell>
 
   //-----------------------------------------------------------------------------------------------
   SM2012SphTree(int, int, int, int, FLOAT, FLOAT, FLOAT, string, string,
-                DomainBox<ndim> *, SmoothingKernel<ndim> *, CodeTiming *);
+                DomainBox<ndim> *, SmoothingKernel<ndim> *, CodeTiming *,
+                ParticleTypeRegister&);
 
 
   //-----------------------------------------------------------------------------------------------
@@ -426,65 +311,143 @@ class SM2012SphTree: public SphTree<ndim,ParticleType,TreeCell>
 
 };
 
-
+//=================================================================================================
+// Tree constructor factory templates
+///
+/// These are simple functions to construct the correct tree based upon the given arguments.
+///
+/// A thin proxy struct is used to ensure that the template specialisation can be done cleanly and
+/// correctly since partial template specialisation is needed. The struct template below should be
+/// specialized as required.
+//=================================================================================================
 
 //=================================================================================================
-//  Class SM2012SphKDTree
-/// \brief   Class containing kd-tree for computing SM2012 SPH force loops.
-/// \details Class containing kd-tree for computing SM2012 SPH force loops.
-/// \author  D. A. Hubber
-/// \date    17/09/2014
+// struct __construct_tree_impl
+// The implementation structs for the tree constructor factory functions.
 //=================================================================================================
-template <int ndim, template<int> class ParticleType, template<int> class TreeCell>
-class SM2012SphKDTree: public SM2012SphTree<ndim,ParticleType,TreeCell>
+template<int ndim, template<int> class ParticleType, template<int> class TreeCell>
+struct __construct_tree_impl
 {
- public:
+	typedef Tree<ndim, ParticleType, TreeCell> return_type ;
 
-  using SphTree<ndim,ParticleType,TreeCell>::tree;
-  using SphTree<ndim,ParticleType,TreeCell>::ghosttree;
-#ifdef MPI_PARALLEL
-  using SphTree<ndim,ParticleType,TreeCell>::mpighosttree;
-  using SphTree<ndim,ParticleType,TreeCell>::Nmpi;
-  using SphTree<ndim,ParticleType,TreeCell>::prunedtree;
-  using SphTree<ndim,ParticleType,TreeCell>::sendprunedtree;
-#endif
-
-
-  //-----------------------------------------------------------------------------------------------
-  SM2012SphKDTree(int, int, int, int, FLOAT, FLOAT, FLOAT, string, string,
-                  DomainBox<ndim> *, SmoothingKernel<ndim> *, CodeTiming *,
-                  ParticleTypeRegister&);
-
+	static return_type*  construct(int Nleafmaxaux, FLOAT thetamaxsqdaux,
+								   FLOAT kernrangeaux, FLOAT macerroraux,
+								   string gravity_mac_aux, string multipole_aux,
+								   const DomainBox<ndim>& domain,
+								   const ParticleTypeRegister& reg)
+	{
+	  string message = "Tree cell type for GradhSphTree not recognised." ;
+	  ExceptionHandler::getIstance().raise(message);
+	  return NULL ;
+	}
 };
-
-
-
 //=================================================================================================
-//  Class SM2012SphOctTree
-/// \brief   Class containing octal-tree for computing SM2012 SPH summation and force loops.
-/// \details Class containing octal-tree for computing SM2012 SPH summation and force loops.
-/// \author  D. A. Hubber
-/// \date    17/09/2014
+//  new_tree
+/// Construct a single tree based on the template types.
+/// This function is used to construct at KD, Oct or BruteForce as required from the TreeCell type
+/// DO NOT EVER SPECIALIZE THIS TEMPLATE. BAD THINGS ARE GUARANTEED TO HAPPEN. DON'T BLAME ME.
+///   USE THE __construct_tree_impl STRUCTURE INSTEAD.
 //=================================================================================================
-template <int ndim, template<int> class ParticleType, template<int> class TreeCell>
-class SM2012SphOctTree: public SM2012SphTree<ndim,ParticleType,TreeCell>
+template<int ndim, template<int> class ParticleType, template<int> class TreeCell>
+Tree<ndim, ParticleType, TreeCell>* new_tree(int Nleafmax, FLOAT thetamaxsqd,
+		   	   	   	   	 	 	 	 	 	 FLOAT kernrange, FLOAT macerror,
+		   	   	   	   	 	 	 	 	 	 string gravity_mac, string multipole,
+		   	   	   	   	 	 	 	 	 	 const DomainBox<ndim>& domain,
+		   	   	   	   	 	 	 	 	 	 const ParticleTypeRegister& reg)
 {
- public:
+  return __construct_tree_impl<ndim, ParticleType, TreeCell>::construct
+		  (Nleafmax, thetamaxsqd, kernrange, macerror,
+		  gravity_mac,  multipole,domain, reg) ;
+}
+//=================================================================================================
+//  new_tree_array
+/// Construct an array of pointer to tree based on the template types.
+/// This function is used to construct at KD, Oct or BruteForce as required from the TreeCell type.
+/// DO NOT EVER SPECIALIZE THIS TEMPLATE. BAD THINGS ARE GUARANTEED TO HAPPEN. DON'T BLAME ME.
+///   USE THE __construct_tree_impl STRUCTURE INSTEAD.
+//=================================================================================================
+template<int ndim, template<int> class ParticleType, template<int> class TreeCell>
+Tree<ndim, ParticleType, TreeCell>** new_tree_array(int NumTrees)
+{
+  typename __construct_tree_impl<ndim, ParticleType, TreeCell>::return_type** derived =
+      __construct_tree_impl<ndim, ParticleType, TreeCell>::construct_array(NumTrees) ;
 
-  using SphTree<ndim,ParticleType,TreeCell>::tree;
-  using SphTree<ndim,ParticleType,TreeCell>::ghosttree;
-#ifdef MPI_PARALLEL
-  using SphTree<ndim,ParticleType,TreeCell>::mpighosttree;
-  using SphTree<ndim,ParticleType,TreeCell>::Nmpi;
-  using SphTree<ndim,ParticleType,TreeCell>::prunedtree;
-  using SphTree<ndim,ParticleType,TreeCell>::sendprunedtree;
-#endif
+  return reinterpret_cast<Tree<ndim,ParticleType,TreeCell>**>(derived) ;
+}
 
 
-  //-----------------------------------------------------------------------------------------------
-  SM2012SphOctTree(int, int, int, int, FLOAT, FLOAT, FLOAT, string, string,
-                   DomainBox<ndim> *, SmoothingKernel<ndim> *, CodeTiming *,
-                   ParticleTypeRegister&);
+//=================================================================================================
+// struct __construct_tree_impl
+// KDTree specialisation
+//=================================================================================================
+template<int ndim, template<int> class ParticleType>
+struct  __construct_tree_impl<ndim, ParticleType, KDTreeCell> {
+  typedef KDTree<ndim,ParticleType, KDTreeCell> return_type ;
 
-};
+  static return_type*  construct(int Nleafmax, FLOAT thetamaxsqd,
+		                         FLOAT kernrange, FLOAT macerror,
+	 	 	 	 	 	         string gravity_mac, string multipole,
+	 	 	 	 	 	         const DomainBox<ndim>& domain,
+	 	 	 	 	 	         const ParticleTypeRegister& reg)
+  {
+	return new return_type(Nleafmax, thetamaxsqd, kernrange, macerror,
+			 	 	 	   gravity_mac,  multipole,domain, reg) ;
+  }
+} ;
+//=================================================================================================
+// struct __construct_tree_impl
+// OctTree specialisation
+//=================================================================================================
+template<int ndim, template<int> class ParticleType>
+struct  __construct_tree_impl<ndim, ParticleType, OctTreeCell> {
+  typedef OctTree<ndim,ParticleType, OctTreeCell> return_type ;
+
+  static return_type*  construct(int Nleafmax, FLOAT thetamaxsqd,
+		                         FLOAT kernrange, FLOAT macerror,
+	 	 	 	 	 	         string gravity_mac, string multipole,
+	 	 	 	 	 	         const DomainBox<ndim>& domain,
+	 	 	 	 	 	         const ParticleTypeRegister& reg)
+  {
+	return new return_type(Nleafmax, thetamaxsqd, kernrange, macerror,
+			 	 	 	   gravity_mac,  multipole,domain, reg) ;
+  }
+} ;
+//=================================================================================================
+// struct __construct_tree_impl
+// OctTree specialisation with TreeRay Cells
+//=================================================================================================
+template<int ndim, template<int> class ParticleType>
+struct  __construct_tree_impl<ndim, ParticleType, TreeRayCell> {
+  typedef OctTree<ndim,ParticleType, TreeRayCell> return_type ;
+
+  static return_type*  construct(int Nleafmax, FLOAT thetamaxsqd,
+		                         FLOAT kernrange, FLOAT macerror,
+	 	 	 	 	 	         string gravity_mac, string multipole,
+	 	 	 	 	 	         const DomainBox<ndim>& domain,
+	 	 	 	 	 	         const ParticleTypeRegister& reg)
+  {
+	return new return_type(Nleafmax, thetamaxsqd, kernrange, macerror,
+			 	 	 	   gravity_mac,  multipole,domain, reg) ;
+  }
+} ;
+//=================================================================================================
+// struct __construct_tree_impl
+// Brute Force Tree specialisation
+//=================================================================================================
+template<int ndim, template<int> class ParticleType>
+struct  __construct_tree_impl<ndim, ParticleType, BruteForceTreeCell> {
+  typedef BruteForceTree<ndim,ParticleType, BruteForceTreeCell> return_type ;
+
+  static return_type*  construct(int Nleafmax, FLOAT thetamaxsqd,
+		                         FLOAT kernrange, FLOAT macerror,
+	 	 	 	 	 	         string gravity_mac, string multipole,
+	 	 	 	 	 	         const DomainBox<ndim>& domain,
+	 	 	 	 	 	         const ParticleTypeRegister& reg)
+  {
+	return new return_type(Nleafmax, thetamaxsqd, kernrange, macerror,
+			 	 	 	   gravity_mac,  multipole,domain, reg) ;
+  }
+} ;
+
+
 #endif
