+++ src/Makefile
@@ -181,11 +181,10 @@ OBJ += M4Kernel.o QuinticKernel.o GaussianKernel.o TabulatedKernel.o
 OBJ += SphIntegration.o SphLeapfrogKDK.o SphLeapfrogDKD.o
 OBJ += RiemannSolver.o
 OBJ += SphNeighbourSearch.o
-OBJ += BruteForceSearch.o HydroTree.o Tree.o KDTree.o OctTree.o
-OBJ += SphBruteForceSearch.o
-OBJ += MeshlessFVBruteForce.o MeshlessFVTree.o
-OBJ += GradhSphBruteForce.o GradhSphTree.o
-OBJ += SM2012SphBruteForce.o SM2012SphTree.o
+OBJ += HydroTree.o Tree.o KDTree.o OctTree.o BruteForceTree.o
+OBJ += MeshlessFVBruteForce.o MeshlessFVTree.o BruteForceSearch.o
+OBJ += GradhSphTree.o
+OBJ += SM2012SphTree.o
 OBJ += Ewald.o
 OBJ += AdiabaticEOS.o IsothermalEOS.o BarotropicEOS.o Barotropic2EOS.o RadwsEOS.o
 OBJ += IonisingRadiationEOS.o MCRadiationEOS.o
@@ -206,7 +205,6 @@ OBJ += CodeTiming.o
 OBJ += IC.o
 OBJ += Dust.o
 OBJ += Particle.o
-OBJ += BruteForceTree.o
 
 
 ifeq ($(MPI),1)
