+++ src/SM2013/SM2012SphSimulation.cpp
@@ -231,16 +231,18 @@ void SM2012SphSimulation<ndim>::ProcessSphParameters(void)
   // Create neighbour searching object based on chosen method in params file
   //-------------------------------------------------------------------------
   if (stringparams["neib_search"] == "bruteforce")
-    sphneib = new SM2012SphBruteForce<ndim,SM2012SphParticle>
-     (sph->kernp->kernrange,&simbox,sph->kernp,this->timing);
+    sphneib = new SM2012SphTree<ndim,SM2012SphParticle,BruteForceTreeCell>
+  	  (intparams["Nleafmax"], Nmpi, intparams["pruning_level_min"], intparams["pruning_level_max"],
+  	   floatparams["thetamaxsqd"], sph->kernp->kernrange, floatparams["macerror"],
+  	   stringparams["gravity_mac"], stringparams["multipole"], &simbox, sph->kernp, timing, sph->types);
   else if (stringparams["neib_search"] == "kdtree") {
-    sphneib = new SM2012SphKDTree<ndim,SM2012SphParticle,KDTreeCell>
+    sphneib = new SM2012SphTree<ndim,SM2012SphParticle,KDTreeCell>
      (intparams["Nleafmax"], Nmpi, intparams["pruning_level_min"], intparams["pruning_level_max"],
       floatparams["thetamaxsqd"], sph->kernp->kernrange, floatparams["macerror"],
       stringparams["gravity_mac"], stringparams["multipole"], &simbox, sph->kernp, timing, sph->types);
   }
   else if (stringparams["neib_search"] == "octtree") {
-    sphneib = new SM2012SphOctTree<ndim,SM2012SphParticle,OctTreeCell>
+    sphneib = new SM2012SphTree<ndim,SM2012SphParticle,OctTreeCell>
      (intparams["Nleafmax"], Nmpi, intparams["pruning_level_min"], intparams["pruning_level_max"],
       floatparams["thetamaxsqd"], sph->kernp->kernrange, floatparams["macerror"],
       stringparams["gravity_mac"], stringparams["multipole"], &simbox, sph->kernp, timing, sph->types);
