+++ src/SM2013/SM2012SphTree.cpp
@@ -40,102 +40,6 @@
 #endif
 using namespace std;
 
-
-
-//=================================================================================================
-//  SM2012SphKDTree::SM2012SphKDTree
-/// SM2012SphKDTree constructor.  Initialises various variables and creates tree objects.
-//=================================================================================================
-template <int ndim, template<int> class ParticleType, template<int> class TreeCell>
-SM2012SphKDTree<ndim,ParticleType,TreeCell>::SM2012SphKDTree
- (int _Nleafmax, int _Nmpi, int _pruning_level_min, int _pruning_level_max, FLOAT _thetamaxsqd,
-  FLOAT _kernrange, FLOAT _macerror, string _gravity_mac, string _multipole,
-  DomainBox<ndim>* _box, SmoothingKernel<ndim>* _kern, CodeTiming* _timing, ParticleTypeRegister& types):
- NeighbourSearch<ndim>(_kernrange, _box, _kern, _timing),
- SM2012SphTree<ndim,ParticleType,TreeCell>
-  (_Nleafmax, _Nmpi, _pruning_level_min, _pruning_level_max, _thetamaxsqd,
-   _kernrange, _macerror, _gravity_mac, _multipole, _box, _kern, _timing)
-{
-  // Set-up main tree object
-  tree = new KDTree<ndim,ParticleType,TreeCell>(_Nleafmax, _thetamaxsqd, _kernrange,
-                                                _macerror, _gravity_mac, _multipole, *_box, types);
-
-  // Set-up ghost-particle tree object
-  ghosttree = new KDTree<ndim,ParticleType,TreeCell>(_Nleafmax, _thetamaxsqd, _kernrange,
-                                                     _macerror, _gravity_mac, _multipole, *_box, types);
-
-#ifdef MPI_PARALLEL
-  // Set-up ghost-particle tree object
-  mpighosttree = new KDTree<ndim,ParticleType,TreeCell>(_Nleafmax, _thetamaxsqd, _kernrange,
-                                                        _macerror, _gravity_mac, _multipole, *_box, types);
-
-  // Set-up multiple pruned trees, one for each MPI process
-  KDTree<ndim,ParticleType,TreeCell>** prunedtree_derived = new KDTree<ndim,ParticleType,TreeCell>*[Nmpi];
-  prunedtree = (Tree<ndim,ParticleType,TreeCell> **) prunedtree_derived;
-  KDTree<ndim,ParticleType,TreeCell>** sendprunedtree_derived = new KDTree<ndim,ParticleType,TreeCell>*[Nmpi];
-  sendprunedtree = (Tree<ndim,ParticleType,TreeCell> **) sendprunedtree_derived;
-
-  for (int i=0; i<Nmpi; i++) {
-    prunedtree[i] = new KDTree<ndim,ParticleType,TreeCell>
-     (_Nleafmax, _thetamaxsqd, _kernrange, _macerror, _gravity_mac, _multipole, *_box, types);
-  }
-  for (int i=0; i<Nmpi; i++) {
-    sendprunedtree[i] = new KDTree<ndim,ParticleType,TreeCell>
-     (_Nleafmax, _thetamaxsqd, _kernrange, _macerror, _gravity_mac, _multipole, *_box, types);
-  }
-#endif
-}
-
-
-
-//=================================================================================================
-//  SM2012SphOctTree::SM2012SphOctTree
-/// SphTree constructor.  Initialises various variables.
-//=================================================================================================
-template <int ndim, template<int> class ParticleType, template<int> class TreeCell>
-SM2012SphOctTree<ndim,ParticleType,TreeCell>::SM2012SphOctTree
- (int _Nleafmax, int _Nmpi, int _pruning_level_min, int _pruning_level_max, FLOAT _thetamaxsqd,
-  FLOAT _kernrange, FLOAT _macerror, string _gravity_mac, string _multipole,
-  DomainBox<ndim>* _box, SmoothingKernel<ndim>* _kern, CodeTiming* _timing, ParticleTypeRegister& types):
- NeighbourSearch<ndim>(_kernrange, _box, _kern, _timing),
- SM2012SphTree<ndim,ParticleType,TreeCell>
-  (_Nleafmax, _Nmpi, _pruning_level_min, _pruning_level_max, _thetamaxsqd,
-   _kernrange, _macerror, _gravity_mac, _multipole, _box, _kern, _timing)
-{
-  // Set-up main tree object
-  tree = new OctTree<ndim,ParticleType,TreeCell>(_Nleafmax, _thetamaxsqd, _kernrange,
-                                                 _macerror, _gravity_mac, _multipole, *_box, types);
-
-  // Set-up ghost-particle tree object
-  ghosttree = new OctTree<ndim,ParticleType,TreeCell>(_Nleafmax, _thetamaxsqd, _kernrange,
-                                                      _macerror, _gravity_mac, _multipole, *_box, types);
-
-#ifdef MPI_PARALLEL
-  // Set-up ghost-particle tree object
-  mpighosttree = new OctTree<ndim,ParticleType,TreeCell>(_Nleafmax, _thetamaxsqd, _kernrange,
-                                                         _macerror, _gravity_mac, _multipole, *_box, types);
-
-  // Set-up multiple pruned trees, one for each MPI process
-  //*(prunedtree) = *(new OctTree<ndim,ParticleType,TreeCell>*[Nmpi]);
-  // Set-up multiple pruned trees, one for each MPI process
-  OctTree<ndim,ParticleType,TreeCell>** prunedtree_derived = new OctTree<ndim,ParticleType,TreeCell>*[Nmpi];
-  prunedtree = (Tree<ndim,ParticleType,TreeCell> **) prunedtree_derived;
-  OctTree<ndim,ParticleType,TreeCell>** sendprunedtree_derived = new OctTree<ndim,ParticleType,TreeCell>*[Nmpi];
-  sendprunedtree = (Tree<ndim,ParticleType,TreeCell> **) sendprunedtree_derived;
-
-  for (int j=0; j<Nmpi; j++) {
-    prunedtree[j] = new OctTree<ndim,ParticleType,TreeCell>
-     (_Nleafmax, _thetamaxsqd, _kernrange, _macerror, _gravity_mac, _multipole, *_box, types);
-  }
-  for (int i=0; i<Nmpi; i++) {
-    sendprunedtree[i] = new OctTree<ndim,ParticleType,TreeCell>
-     (_Nleafmax, _thetamaxsqd, _kernrange, _macerror, _gravity_mac, _multipole, *_box, types);
-  }
-#endif
-}
-
-
-
 //=================================================================================================
 //  SM2012SphTree::SM2012SphTree
 /// SM2012SphTree constructor.  Initialises various variables.
@@ -144,41 +48,52 @@ template <int ndim, template<int> class ParticleType, template<int> class TreeCe
 SM2012SphTree<ndim,ParticleType,TreeCell>::SM2012SphTree
  (int _Nleafmax, int _Nmpi, int _pruning_level_min, int _pruning_level_max, FLOAT _thetamaxsqd,
   FLOAT _kernrange, FLOAT _macerror, string _gravity_mac, string _multipole,
-  DomainBox<ndim>* _box, SmoothingKernel<ndim>* _kern, CodeTiming* _timing):
+  DomainBox<ndim>* _box, SmoothingKernel<ndim>* _kern, CodeTiming* _timing, ParticleTypeRegister& types):
  NeighbourSearch<ndim>(_kernrange, _box, _kern, _timing),
  SphTree<ndim,ParticleType,TreeCell>
   (_Nleafmax, _Nmpi, _pruning_level_min, _pruning_level_max, _thetamaxsqd,
    _kernrange, _macerror, _gravity_mac, _multipole, _box, _kern, _timing)
 {
+  // Set-up main tree object
+  tree = new_tree<ndim,ParticleType,TreeCell>(_Nleafmax, _thetamaxsqd, _kernrange,
+   		  	  	  	  	  	  	  	  	  	  _macerror, _gravity_mac, _multipole, *_box, types);
+
+  // Set-up ghost-particle tree object
+  ghosttree = new_tree<ndim,ParticleType,TreeCell>(_Nleafmax, _thetamaxsqd, _kernrange,
+		  _macerror, _gravity_mac, _multipole, *_box, types);
+
+#ifdef MPI_PARALLEL
+  // Set-up ghost-particle tree object
+  mpighosttree = new_tree<ndim,ParticleType,TreeCell>(_Nleafmax, _thetamaxsqd, _kernrange,
+		  _macerror, _gravity_mac, _multipole, *_box, types);
+
+  // Set-up multiple pruned trees, one for each MPI process
+  prunedtree = new Tree<ndim,ParticleType,TreeCell>*[Nmpi] ;
+  // new_tree_array<ndim,ParticleType,TreeCell>(Nmpi);
+  sendprunedtree =  new Tree<ndim,ParticleType,TreeCell>*[Nmpi] ;
+  //new_tree_array<ndim,ParticleType,TreeCell>(Nmpi);
+
+  for (int i=0; i<Nmpi; i++) {
+	prunedtree[i] = new_tree<ndim,ParticleType,TreeCell>(_Nleafmax, _thetamaxsqd, _kernrange,
+   	  	  	  	  	  	  	  	  	  	  	  	  	  	 _macerror, _gravity_mac, _multipole, *_box, types);
+  }
+  for (int i=0; i<Nmpi; i++) {
+	sendprunedtree[i] = new_tree<ndim,ParticleType,TreeCell>(_Nleafmax, _thetamaxsqd, _kernrange,
+       														 _macerror, _gravity_mac, _multipole, *_box, types);
+  }
+#endif
+
 }
 
 
 
-//=================================================================================================
-//  SM2012SphTree::~SM2012SphTree
-/// SM2012SphTree destructor.  Deallocates tree memory upon object destruction.
-//=================================================================================================
-/*template <int ndim, template<int> class ParticleType, template<int> class TreeCell>
-SM2012SphTree<ndim,ParticleType,TreeCell>::~SM2012SphTree()
-{
-  if (tree->allocated_tree) {
-    this->DeallocateMemory();
-    tree->DeallocateTreeMemory();
-  }
-}*/
-
-
-
+template class SM2012SphTree<1,SM2012SphParticle,BruteForceTreeCell>;
+template class SM2012SphTree<2,SM2012SphParticle,BruteForceTreeCell>;
+template class SM2012SphTree<3,SM2012SphParticle,BruteForceTreeCell>;
 template class SM2012SphTree<1,SM2012SphParticle,KDTreeCell>;
 template class SM2012SphTree<2,SM2012SphParticle,KDTreeCell>;
 template class SM2012SphTree<3,SM2012SphParticle,KDTreeCell>;
-template class SM2012SphKDTree<1,SM2012SphParticle,KDTreeCell>;
-template class SM2012SphKDTree<2,SM2012SphParticle,KDTreeCell>;
-template class SM2012SphKDTree<3,SM2012SphParticle,KDTreeCell>;
-
 template class SM2012SphTree<1,SM2012SphParticle,OctTreeCell>;
 template class SM2012SphTree<2,SM2012SphParticle,OctTreeCell>;
 template class SM2012SphTree<3,SM2012SphParticle,OctTreeCell>;
-template class SM2012SphOctTree<1,SM2012SphParticle,OctTreeCell>;
-template class SM2012SphOctTree<2,SM2012SphParticle,OctTreeCell>;
-template class SM2012SphOctTree<3,SM2012SphParticle,OctTreeCell>;
+
